requirements.txt
Flask>=2.0
pdfminer.six>=20221105
python-dotenv>=1.0
requests>=2.28
click>=8.0
pytest>=7.0
reportlab>=4.0
